refactor: 인증 및 인가, JWT 관련 클래스들을 하나의 패키지로 통합
