+++ src/main/java/wad/seoul_nolgoat/auth/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.service.auth;
+package wad.seoul_nolgoat.auth;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -7,10 +7,10 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
+import wad.seoul_nolgoat.auth.dto.*;
 import wad.seoul_nolgoat.domain.user.User;
 import wad.seoul_nolgoat.domain.user.UserRepository;
 import wad.seoul_nolgoat.exception.ApiException;
-import wad.seoul_nolgoat.service.auth.dto.*;
 
 import static wad.seoul_nolgoat.exception.ErrorCode.UNSUPPORTED_PROVIDER;
 import static wad.seoul_nolgoat.exception.ErrorCode.USER_NOT_FOUND;
