+++ src/main/java/wad/seoul_nolgoat/auth/CustomSuccessHandler.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.service.auth;
+package wad.seoul_nolgoat.auth;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,8 +8,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
-import wad.seoul_nolgoat.jwt.JwtService;
-import wad.seoul_nolgoat.service.auth.dto.OAuth2UserImpl;
+import wad.seoul_nolgoat.auth.dto.OAuth2UserImpl;
+import wad.seoul_nolgoat.auth.jwt.JwtService;
 
 import java.io.IOException;
 import java.net.URLEncoder;
