+++ src/main/java/wad/seoul_nolgoat/auth/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.service.auth.dto;
+package wad.seoul_nolgoat.auth.dto;
 
 public interface OAuth2Response {
