+++ src/main/java/wad/seoul_nolgoat/config/SecurityConfig.java
@@ -14,11 +14,11 @@ import org.springframework.security.web.authentication.logout.LogoutFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import wad.seoul_nolgoat.jwt.JwtFilter;
-import wad.seoul_nolgoat.jwt.JwtService;
-import wad.seoul_nolgoat.service.auth.CustomLogoutFilter;
-import wad.seoul_nolgoat.service.auth.CustomOAuth2UserService;
-import wad.seoul_nolgoat.service.auth.CustomSuccessHandler;
+import wad.seoul_nolgoat.auth.CustomLogoutFilter;
+import wad.seoul_nolgoat.auth.CustomOAuth2UserService;
+import wad.seoul_nolgoat.auth.CustomSuccessHandler;
+import wad.seoul_nolgoat.auth.jwt.JwtFilter;
+import wad.seoul_nolgoat.auth.jwt.JwtService;
 
 import java.util.List;
